Safari-611.1.10/LayoutTests/fast/forms/validation-custom-message-expected.txt

Required text input:
Tests that the custom validation message set by JavaScript is displayed on UI side.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS message is "This is a custom validity message."
PASS message is "Fill out this field"
PASS successfullyParsed is true

TEST COMPLETE
